+++ shift/system/config/default.php
@@ -30,8 +30,8 @@ $_['database'] = [
 ];
 
 $_['session'] = [
-    'name'     => 'SHIFTID',
-    'use_cookies'      => '1',
-    'use_trans_sid'    => '0',
-    'sid_length'       => rand(48, 64),
+    'session_name'  => 'SHIFTID',
+    'use_cookies'   => '1',
+    'use_trans_sid' => '0',
+    'sid_length'    => rand(48, 64),
 ];